requirements.txt
fastapi
uvicorn
requests
python-dotenv
supabase
openai
PyMuPDF
